Average score and median score are now updated in the backend. Made the necessary adjustments to the post quiz page so that it utilizes the new data from the backend

+++ server/graphql/typeDefs/quizattempts-def.js
@@ -6,7 +6,7 @@ module.exports = gql`
         user: User
         quiz: Quiz
         elapsedTime: String
-        score: Int!
+        score: Float!
         answerChoices: [[String!]!]!
         questions: [String!]!
         numCorrect: Int!
